+++ src/include/poac/util.hpp
@@ -1,0 +1,7 @@
+#ifndef POAC_UTIL_HPP
+#define POAC_UTIL_HPP
+
+#include "util/file.hpp"
+#include "util/requests.hpp"
+
+#endif // !POAC_UTIL_HPP
