--- src/include/poac/utility.hpp
@@ -1,7 +1,0 @@
-#ifndef POAC_UTILITY_HPP
-#define POAC_UTILITY_HPP
-
-#include "utility/file.hpp"
-#include "utility/requests.hpp"
-
-#endif // !POAC_UTILITY_HPP
